# Documentation/dev-guide/apispec/swagger/rpc.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "etcdserver/etcdserverpb/rpc.proto",
    "version": "version not set"
  },
  "paths": {
...
    "/v3beta/maintenance/snapshot": {
      "post": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Snapshot sends a snapshot of the entire backend from a member over a stream to a client.",
        "operationId": "Snapshot",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbSnapshotRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/etcdserverpbSnapshotResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "etcdserverpbResponseHeader": {
      "type": "object",
      "properties": {
        "cluster_id": {
          "description": "cluster_id is the ID of the cluster which sent the response.",
          "type": "string",
          "format": "uint64"
        },
        "member_id": {
          "description": "member_id is the ID of the member which sent the response.",
          "type": "string",
          "format": "uint64"
        },
        "raft_term": {
          "description": "raft_term is the raft term when the request was applied.",
          "type": "string",
          "format": "uint64"
        },
        "revision": {
          "description": "revision is the key-value store revision when the request was applied.",
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "etcdserverpbSnapshotRequest": {
      "type": "object"
    },
    "etcdserverpbSnapshotResponse": {
      "type": "object",
      "properties": {
        "blob": {
          "description": "blob contains the next chunk of the snapshot in the snapshot stream.",
          "type": "string",
          "format": "byte"
        },
        "header": {
          "description": "header has the current key-value store information. The first header in the snapshot\nstream indicates the point in time of the snapshot.",
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        },
        "remaining_bytes": {
          "type": "string",
          "format": "uint64",
          "title": "remaining_bytes is the number of blob bytes to be sent after this message"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKey": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ]
}